Implemented navigation between scenes; created a homescreen scene
